enable mono test ci
