+++ .github/actions/godot-install/action.yml
@@ -60,23 +60,22 @@ runs:
         GODOT_BIN=Godot_v${{ inputs.godot-version }}-${{ inputs.godot-status-version }}_${{ inputs.godot-bin-name }}
         if ${{inputs.godot-mono == 'true'}}; then
           GODOT_BIN=Godot_v${{ inputs.godot-version }}-${{ inputs.godot-status-version }}_mono_${{ inputs.godot-bin-name }}
-          DOWNLOAD_URL=$DOWNLOAD_URL/mono
         fi
 
         GODOT_PACKAGE=$GODOT_BIN.zip
         wget $DOWNLOAD_URL/$GODOT_PACKAGE -P ${{ inputs.godot-cache-path }}
         unzip ${{ inputs.godot-cache-path }}/$GODOT_PACKAGE -d ${{ inputs.godot-cache-path }}
+        rm -rf ${{ inputs.godot-cache-path }}/$GODOT_PACKAGE
 
         if ${{runner.OS == 'Linux'}}; then
           echo "Run linux part"
+
           if ${{inputs.godot-mono == 'true'}}; then
-            mv ${{ inputs.godot-cache-path }}/$GODOT_BIN/* ${{ inputs.godot-cache-path }}/
-            rm -rf ${{ inputs.godot-cache-path }}/$GODOT_BIN*
-            mv ${{ inputs.godot-cache-path }}/Godot_v${{ inputs.godot-version }}* ${{ inputs.godot-cache-path }}/godot
-          else
-            mv ${{ inputs.godot-cache-path }}/$GODOT_BIN ${{ inputs.godot-cache-path }}/godot
+            mv ${{ inputs.godot-cache-path }}/$GODOT_BIN/* ${{ inputs.godot-cache-path }}
+            rmdir ${{ inputs.godot-cache-path }}/$GODOT_BIN/
           fi
 
+          mv ${{ inputs.godot-cache-path }}/Godot_v* ${{ inputs.godot-cache-path }}/godot
           chmod u+x ${{ inputs.godot-cache-path }}/godot
           echo "${{ inputs.godot-cache-path }}/godot"
         else
