+++ .github/workflows/ci-dev.yml
@@ -21,7 +21,7 @@ jobs:
       fail-fast: false
       max-parallel: 10
       matrix:
-        godot-version: ['4.0.1', '4.0.2', '4.0.3', '4.1', '4.1.1', '4.1.2']
+        godot-version: ['4.0.1', '4.0.2', '4.0.3', '4.1', '4.1.1', '4.1.2', '4.1.3']
 
     name: "CI on Godot 🐧 v${{ matrix.godot-version }}-stable"
     uses: ./.github/workflows/unit-tests.yml
