+++ .github/workflows/ci-pr.yml
@@ -22,11 +22,11 @@ jobs:
       fail-fast: false
       max-parallel: 10
       matrix:
-        godot-version: ['4.0.1', '4.0.2', '4.0.3', '4.1', '4.1.1', '4.1.2']
+        godot-version: ['4.0.1', '4.0.2', '4.0.3', '4.1', '4.1.1', '4.1.2', '4.1.3']
         godot-status: ['stable']
         include:
           - godot-version: '4.2'
-            godot-status: 'beta2'
+            godot-status: 'beta4'
 
 
     name: "CI on Godot 🐧 v${{ matrix.godot-version }}-${{ matrix.godot-status }}"
